Add support for the IEI NOVA-4899R 5.25 SBC mainboard (patch submitted by Luis Correia <luis.f.correia@gmail.com>). The code is loosely based on the Eaglelion 5bcm mainboard.
Warning: this is work in progress!

As of now, it does boot with serial console only (no vga), and two ethernet
cards work sometimes. This has to do with the IRQ assignments, which are a
complete mess. USB is now apparently working, but I can't make any device
to be recognized.

The PCI slot is still unusable due to the IRQ thing.

Audio, other serial ports, irda, floppy and paralell port support is
unknown aka untested yet.

(closes #32)

Signed-off-by: Luis Correia <luis.f.correia@gmail.com>
Acked-by: Acked-by: Stefan Reinauer <stepan@coresystems.de>
Acked-by: Uwe Hermann <uwe@hermann-uwe.de>

git-svn-id: svn://svn.coreboot.org/coreboot/trunk@2508 2b7e53f0-3cfb-0310-b3e9-8179ed1497e1