Drop the unfinished, non-working Bitworks IMS board.
It never worked in v2 (the v1 port did work AFAIK, though), and it's
not really useful as reference for other boards anymore (as we now
have a dozen or so 440BX boards which work in v2).

This is a specialized, custom board (not sold on the "public market"),
so it's probably not useful for pretty much everyone out there anyway.

We can easily re-add it later (based on one of the other 440BX boards)
should there be interest and/or someone with the hardware to test.

Signed-off-by: Uwe Hermann <uwe@hermann-uwe.de>
Acked-by: Jordan Crouse <jordan.crouse@amd.com>

git-svn-id: svn://svn.coreboot.org/coreboot/trunk@2988 2b7e53f0-3cfb-0310-b3e9-8179ed1497e1