Temporarily disable the fan control patch from this morning; it turns out to stop the CPU fan on the m57sli v1.1 (PLCC) entirely, which is less than desirable. I did not notice before because my board ran fine for about 15 minutes before the CPU overheated.
Thankfully the board has a good failsafe mode - it just switches off when the
CPU gets too hot, without permanent damage.

I'm debugging this and plan to commit a proper fix later in the week.

This is not really trivial, but the tree is dangerous in the current state so
I'm self-acking.

Signed-off-by: Ward Vandewege <ward@gnu.org>
Acked-by: Ward Vandewege <ward@gnu.org>

git-svn-id: svn://svn.coreboot.org/coreboot/trunk@3118 2b7e53f0-3cfb-0310-b3e9-8179ed1497e1