Doing another 'make' after a 'make clean' was broken until now. Fix it by not deleting build/config.h during 'make clean' (only in 'make distclean').
Also, change the default behaviour of 'make' from asking the user to
run 'make config' (or similar) to actually _run_ 'make config' without
asking questions. It's always possible to explicitly invoke
'make menuconfig' or 'make xconfig' and so on, of course.

Finally, make _all_ targets (allyesconfig, randconfig, and so on)
generate a build/config.h file, as we always #include it.

Signed-off-by: Uwe Hermann <uwe@hermann-uwe.de>
Acked-by: Jordan Crouse <jordan.crouse@amd.com>

git-svn-id: svn://svn.coreboot.org/coreboot/trunk@3211 2b7e53f0-3cfb-0310-b3e9-8179ed1497e1