This is the beginning of support for Geode and Kconfig in v2.
It also brings in the vsm from v3, which was a much cleaner cut.

Over time, I hope to bring all the code back from v3. I have
some rumbas at home and want to use them.

I have a patch which comes in next that makes the rumba build.

Note that I am holding the src/*/amd/Kconfig patch until these get merged.
These have no impact on the current system.

Note that this is not complete but I want to fill in the blanks bit
by bit.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>
Acked-by: Myles Watson <mylesgw@gmail.com>

git-svn-id: svn://svn.coreboot.org/coreboot/trunk@4611 2b7e53f0-3cfb-0310-b3e9-8179ed1497e1